+++ rentACarSpring/src/main/java/com/etiya/rentACarSpring/businnes/concretes/CityManager.java
@@ -56,6 +56,11 @@ public class CityManager implements CityService {
 
     @Override
     public Result update(UpdateCityRequest updateCityRequest) {
+        Result result = BusinnessRules.run(checkIfCityExists(updateCityRequest.getCityId())
+        );
+        if (result != null) {
+            return result;
+        }
         City city = modelMapperService.forRequest().map(updateCityRequest, City.class);
         this.cityDao.save(city);
         return new SuccesResult(Messages.updatedCity);
@@ -63,6 +68,11 @@ public class CityManager implements CityService {
 
     @Override
     public Result delete(DeleteCityRequest deleteCityRequest) {
+        Result result = BusinnessRules.run(checkIfCityExists(deleteCityRequest.getCityId())
+        );
+        if (result != null) {
+            return result;
+        }
         this.cityDao.deleteById(deleteCityRequest.getCityId());
         return new SuccesResult(Messages.deletedCity);
     }
@@ -76,6 +86,14 @@ public class CityManager implements CityService {
         City city = this.cityDao.getByCityName(cityName);
         if (city != null) {
             return new ErrorResult("Şehir kayıtlı.");
+        }
+        return new SuccesResult();
+    }
+
+    @Override
+    public Result checkIfCityExists(int cityId) {
+        if (!this.cityDao.existsById(cityId)) {
+            return new ErrorResult("cityId mevcut değil");
         }
         return new SuccesResult();
     }
