+++ rentACarSpring/src/main/java/com/etiya/rentACarSpring/businnes/concretes/RentalAdditionalServiceManager.java
@@ -5,11 +5,9 @@ import com.etiya.rentACarSpring.businnes.dtos.RentalAdditionalServiceSearchListD
 import com.etiya.rentACarSpring.businnes.request.RentalAdditionalServiceRequest.CreateRentalAdditionalServiceRequest;
 import com.etiya.rentACarSpring.businnes.request.RentalAdditionalServiceRequest.DeleteRentalAdditionalServiceRequest;
 import com.etiya.rentACarSpring.businnes.request.RentalAdditionalServiceRequest.UpdateRentalAdditionalServiceRequest;
+import com.etiya.rentACarSpring.core.utilities.businnessRules.BusinnessRules;
 import com.etiya.rentACarSpring.core.utilities.mapping.ModelMapperService;
-import com.etiya.rentACarSpring.core.utilities.results.DataResult;
-import com.etiya.rentACarSpring.core.utilities.results.Result;
-import com.etiya.rentACarSpring.core.utilities.results.SuccesDataResult;
-import com.etiya.rentACarSpring.core.utilities.results.SuccesResult;
+import com.etiya.rentACarSpring.core.utilities.results.*;
 import com.etiya.rentACarSpring.dataAccess.abstracts.RentalAdditionalServiceDao;
 import com.etiya.rentACarSpring.entities.RentalAdditionalService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -47,6 +45,11 @@ public class RentalAdditionalServiceManager implements RentalAdditionalServiceSe
 
     @Override
     public Result update(UpdateRentalAdditionalServiceRequest updateRentalAdditionalServiceRequest) {
+        Result result = BusinnessRules.run(checkIfRentalAdditionalExists(updateRentalAdditionalServiceRequest.getRentalAdditionalServiceId())
+                );
+        if (result != null) {
+            return result;
+        }
         RentalAdditionalService rentalAdditionalService = modelMapperService.forRequest().map(updateRentalAdditionalServiceRequest, RentalAdditionalService.class);
         this.rentalAdditionalServiceDao.save(rentalAdditionalService);
         return new SuccesResult("Eklendi");
@@ -54,7 +57,20 @@ public class RentalAdditionalServiceManager implements RentalAdditionalServiceSe
 
     @Override
     public Result delete(DeleteRentalAdditionalServiceRequest deleteRentalAdditionalServiceRequest) {
+        Result result = BusinnessRules.run(checkIfRentalAdditionalExists(deleteRentalAdditionalServiceRequest.getRentalAdditionalServiceId())
+        );
+        if (result != null) {
+            return result;
+        }
         this.rentalAdditionalServiceDao.deleteById(deleteRentalAdditionalServiceRequest.getRentalAdditionalServiceId());
         return new SuccesResult("Silindi");
+    }
+
+    @Override
+    public Result checkIfRentalAdditionalExists(int rentalAdditionalId) {
+        if (!this.rentalAdditionalServiceDao.existsById(rentalAdditionalId)) {
+            return new ErrorResult("rentalAdditionalId mevcut değil");
+        }
+        return new SuccesResult();
     }
 }
